Lines Printed Backwards
This task proved difficult. First, I approached the addition of adding a number to the beggining of each line of the poem. This was thought of indexing each line, a string, and using an insert.() at 0 to add a number. However, there was an issue between string and int, stating cannot calcatenate. I went through Stack Overflow, and found an alternative, using enumerate on the strings to yield the number at the beginning of each line, followed by looping it in a forloop until the end of the line. Printing the text backwards afterwards was simple, using the reveresed() command. I as well added an input variable for the user to choose their poem.